Initial state of project
Dependencies are not all dealt with. The target directory requires subdirectories lazysizes-gh-pages and wp-lazysizes-master to be incorporated by hand. Configuration of Apache is not yet documented. Directory structure is not yet documented. The daring can try it anyway by putting the project directory on the drive in which their photos are kept and configuring Apache to serve the PARENT directory as the home directory "pics". Better description will follow.